Renamed Main classes to 'ClientMain' and 'ServerMain'
