+++ client/pom.xml
@@ -21,7 +21,7 @@
                 <artifactId>exec-maven-plugin</artifactId>
                 <version>3.1.1</version>
                 <configuration>
-                    <mainClass>client.Main</mainClass>
+                    <mainClass>client.ClientMain</mainClass>
                 </configuration>
             </plugin>
             <plugin>
@@ -38,7 +38,7 @@
                             <archive>
                                 <manifest>
                                     <mainClass>
-                                        client.Main
+                                        client.ClientMain
                                     </mainClass>
                                 </manifest>
                             </archive>
