+++ client/src/main/java/client/ClientMain.java
@@ -1,10 +1,10 @@
-package server;
+package client;
 
 import chess.*;
 
-public class Main {
+public class ClientMain {
     public static void main(String[] args) {
         var piece = new ChessPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.PAWN);
-        System.out.println("♕ 240 Chess Server: " + piece);
+        System.out.println("♕ 240 Chess Client: " + piece);
     }
 }
