+++ server/src/main/java/server/ServerMain.java
@@ -1,10 +1,10 @@
-package client;
+package server;
 
 import chess.*;
 
-public class Main {
+public class ServerMain {
     public static void main(String[] args) {
         var piece = new ChessPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.PAWN);
-        System.out.println("♕ 240 Chess Client: " + piece);
+        System.out.println("♕ 240 Chess Server: " + piece);
     }
 }
